Refactoring: reorder functions & sourceNode -> astNode
